Alarm funzionano solo ad app aperta, però ci sto lavorando

+++ app/src/main/java/com/uninsubria/pillskeeper/Farmaco.kt
@@ -3,8 +3,8 @@ package com.uninsubria.pillskeeper
 import com.google.firebase.storage.FirebaseStorage
 import com.google.firebase.storage.StorageReference
 
-data class Farmaco(var name:String, var mImageUrl:String, var qTot: Int, var q: Int, var every: String, var time: String) : java.io.Serializable {//serializable serve per putExtra
-    constructor() : this("","",0,0,"","")
+data class Farmaco(var name:String, var mImageUrl:String, var qTot: Int, var q: Double, var every: String, var time: String) : java.io.Serializable {//serializable serve per putExtra
+    constructor() : this("","",0, 0.0,"","")
 
     fun convertImg(img: String = mImageUrl ): StorageReference {
         return FirebaseStorage.getInstance().getReferenceFromUrl("gs://prove-b822e.appspot.com$img")
